Bug of missing face resolved, plus correction of types, regarding to the protocol of Mr Lauri.

+++ Template/src/diamond.cpp
@@ -6,7 +6,7 @@ Diamond::Diamond()
 
 }
 
-Diamond::Diamond(float pavillon, float crown, float rondiste, float table, float radius, float complexity, const GLfloat color[3], float alpha, float lvlPavillon, float lvlCrown)
+Diamond::Diamond(float32 pavillon, float32 crown, float32 rondiste, float32 table, float32 radius, int32 complexity, const GLfloat color[3], float32 alpha, float32 lvlPavillon, float32 lvlCrown)
 {
     //complexity must be a multiple of 2
     if(fmod(complexity,2) != 0)
